App/requirements.txt
python-dotenv
firebase_admin
flask
pymongo
flask_bcrypt
mysql-connector-python
PyMySQL

